backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
gunicorn==21.2.0
pytest==7.4.0
requests==2.31.0
cryptography==41.0.3
